# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OCPCTL is a self-service OpenShift cluster provisioning and management platform. This API provides endpoints for creating, managing, and destroying OpenShift clusters with automated lifecycle management, policy enforcement, and resource tracking.",
        "title": "OCPCTL API",
        "termsOfService": "http://github.com/tsanders-rh/ocpctl",
        "contact": {
            "name": "OCPCTL Support",
            "url": "http://github.com/tsanders-rh/ocpctl/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clusters/{id}/kubeconfig": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Downloads the kubeconfig file for a ready cluster as a YAML attachment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/x-yaml"
                ],
                "tags": [
                    "Clusters"
                ],
                "summary": "Download kubeconfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kubeconfig YAML file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Cluster not ready",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not cluster owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Cluster or kubeconfig not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clusters/{id}/kubeconfig/download-url": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a pre-signed S3 URL for downloading the kubeconfig (15-minute expiration)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clusters"
                ],
                "summary": "Get kubeconfig download URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contains download_url field",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Cluster not ready or S3 storage not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not cluster owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Cluster not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clusters/{id}/outputs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns cluster outputs including kubeconfig, credentials, API URLs, and console URL. Only available for ready clusters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clusters"
                ],
                "summary": "Get cluster outputs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ClusterOutputsResponse"
                        }
                    },
                    "400": {
                        "description": "Cluster not ready or outputs not available",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not cluster owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Cluster not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.ClusterOutputsResponse": {
            "type": "object",
            "properties": {
                "api_url": {
                    "type": "string"
                },
                "cluster_id": {
                    "type": "string"
                },
                "cluster_name": {
                    "type": "string"
                },
                "console_url": {
                    "type": "string"
                },
                "kubeadmin": {
                    "description": "Actual credentials",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_api.KubeadminCredentials"
                        }
                    ]
                },
                "kubeadmin_secret_ref": {
                    "description": "Reference to secret location",
                    "type": "string"
                },
                "kubeconfig": {
                    "description": "Full kubeconfig content",
                    "type": "string"
                },
                "kubeconfig_s3_uri": {
                    "description": "S3 URI to kubeconfig file",
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.KubeadminCredentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Format: \"Bearer {token}\". Obtain from /api/v1/auth/login endpoint.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}